+++ lib/src/metric/regression/mape.dart
@@ -1,3 +1,4 @@
+import 'package:ml_algo/src/common/exception/matrix_column_exception.dart';
 import 'package:ml_algo/src/metric/metric.dart';
 import 'package:ml_linalg/linalg.dart';
 
@@ -5,15 +6,20 @@ class MapeMetric implements Metric {
   const MapeMetric();
 
   @override
-  double getScore(Matrix predictedLabels, Matrix origLabels) {
-    if (predictedLabels.columnsNum != 1 || origLabels.columnsNum != 1) {
-      throw Exception('Both predicted labels and original labels have to be '
-          'a matrix-column');
+  double getScore(Matrix predictedLabels, Matrix originalLabels) {
+    if (predictedLabels.columnsNum != 1) {
+      throw MatrixColumnException(predictedLabels);
     }
-    final predicted = predictedLabels.getColumn(0);
-    final original = origLabels.getColumn(0);
 
-    return 100 / predicted.length *
-        ((original - predicted) / original).abs().sum();
+    if (originalLabels.columnsNum != 1) {
+      throw MatrixColumnException(originalLabels);
+    }
+
+    final predicted = predictedLabels.getColumn(0);
+    final original = originalLabels.getColumn(0);
+
+    return ((original - predicted) / original)
+        .abs()
+        .mean();
   }
 }